Extract nameAndDateOfBirth to config/pages
